Project/RAG - Chatbot/requirements.txt
streamlit==1.49.1
python-dotenv==1.0.1
requests==2.32.5

# LangChain ecosystem
langchain
langchain-community==0.4.0
langchain-experimental==0.4.0
langchain-google-genai==1.0.10
langchain-google-community==2.0.2

# LangGraph
langgraph==0.6.7
langgraph-checkpoint-sqlite==2.0.11

# Search utility
duckduckgo-search==6.2.4
